# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API backend be_latihan menggunakan Golang Fiber, GORM, PostgreSQL, dan JWT.",
        "title": "API Praktikum 13 - be_latihan",
        "contact": {
            "name": "Praktikum Pemrograman III",
            "email": "praktikum@example.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/",
    "paths": {
...
        "/api/mahasiswa/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil seluruh data mahasiswa. Endpoint ini membutuhkan token JWT, tetapi tidak membatasi role admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Ambil semua data mahasiswa",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil semua data mahasiswa",
                        "schema": {
                            "$ref": "#/definitions/model.Response200"
                        }
                    },
                    "401": {
                        "description": "Token JWT tidak ada atau tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.Response401"
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan pada server",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menambahkan data mahasiswa baru. Endpoint ini hanya dapat diakses oleh role admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Tambah data mahasiswa",
                "parameters": [
                    {
                        "description": "Payload data mahasiswa",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Mahasiswa"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Mahasiswa berhasil ditambahkan",
                        "schema": {
                            "$ref": "#/definitions/model.Response201"
                        }
                    },
                    "400": {
                        "description": "Payload tidak valid atau NPM/Nama kosong",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "401": {
                        "description": "Token JWT tidak ada atau tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.Response401"
                        }
                    },
                    "403": {
                        "description": "Role user tidak memiliki akses ke endpoint ini",
                        "schema": {
                            "$ref": "#/definitions/model.Response403"
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan pada server",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/mahasiswa/{npm}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil satu data mahasiswa berdasarkan NPM. Endpoint ini hanya dapat diakses oleh role admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Ambil data mahasiswa berdasarkan NPM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NPM mahasiswa",
                        "name": "npm",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data mahasiswa berhasil ditemukan",
                        "schema": {
                            "$ref": "#/definitions/model.Response200"
                        }
                    },
                    "400": {
                        "description": "Format NPM tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "401": {
                        "description": "Token JWT tidak ada atau tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.Response401"
                        }
                    },
                    "403": {
                        "description": "Role user tidak memiliki akses ke endpoint ini",
                        "schema": {
                            "$ref": "#/definitions/model.Response403"
                        }
                    },
                    "404": {
                        "description": "Data mahasiswa tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan pada server",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah data mahasiswa berdasarkan NPM. NPM pada body akan dipaksa mengikuti NPM pada URL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Ubah data mahasiswa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NPM mahasiswa",
                        "name": "npm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload data mahasiswa",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Mahasiswa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data mahasiswa berhasil diubah",
                        "schema": {
                            "$ref": "#/definitions/model.Response200"
                        }
                    },
                    "400": {
                        "description": "Format NPM atau payload tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "401": {
                        "description": "Token JWT tidak ada atau tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.Response401"
                        }
                    },
                    "403": {
                        "description": "Role user tidak memiliki akses ke endpoint ini",
                        "schema": {
                            "$ref": "#/definitions/model.Response403"
                        }
                    },
                    "404": {
                        "description": "Data mahasiswa tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan pada server",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus data mahasiswa berdasarkan NPM. Endpoint ini hanya dapat diakses oleh role admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Hapus data mahasiswa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NPM mahasiswa",
                        "name": "npm",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mahasiswa berhasil dihapus",
                        "schema": {
                            "$ref": "#/definitions/model.Response200"
                        }
                    },
                    "400": {
                        "description": "Format NPM tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "401": {
                        "description": "Token JWT tidak ada atau tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.Response401"
                        }
                    },
                    "403": {
                        "description": "Role user tidak memiliki akses ke endpoint ini",
                        "schema": {
                            "$ref": "#/definitions/model.Response403"
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan pada server",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Mahasiswa": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string",
                    "example": "Bandung"
                },
                "email": {
                    "type": "string",
                    "example": "zidan@example.com"
                },
                "nama": {
                    "type": "string",
                    "example": "Zidan Hairra Ramadhan"
                },
                "no_hp": {
                    "type": "string",
                    "example": "081234567890"
                },
                "npm": {
                    "type": "string",
                    "example": "714240061"
                },
                "prodi": {
                    "type": "string",
                    "example": "D4 Teknik Informatika"
                }
            }
        },
        "model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": "detail error"
                },
                "message": {
                    "type": "string",
                    "example": "detail pesan"
                }
            }
        },
        "model.Response200": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "berhasil mengambil data"
                }
            }
        },
        "model.Response201": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "data berhasil ditambahkan"
                }
            }
        },
        "model.Response401": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "token tidak valid atau tidak ditemukan"
                }
            }
        },
        "model.Response403": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "user tidak memiliki akses untuk fitur ini"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}